requirements.txt
streamlit
python-dotenv
llama-index>=0.10.0
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-faiss
beautifulsoup4
faiss-cpu